feat: initial version of map for service journeys
