+++ src/screens/TripDetails/Map/index.tsx
@@ -4,7 +4,7 @@ import {StyleSheet, View} from 'react-native';
 import {useGeolocationState} from '../../../GeolocationContext';
 import {RouteProp} from '@react-navigation/native';
 import {DetailsStackParams} from '..';
-import {getMapBounds, legsToMapLines, pointOf} from './utils';
+import {createMapLines, getMapBounds, legsToMapLines, pointOf} from './utils';
 import MapRoute from './MapRoute';
 import MapLabel from './MapLabel';
 import {
@@ -15,13 +15,17 @@ import {
   BackArrow,
   useControlPositionsStyle,
 } from '../../../components/map';
-import {Leg} from '../../../sdk';
+import {Leg, MapLeg} from '../../../sdk';
 import Bugsnag from '@bugsnag/react-native';
 import {useTranslation, MapTexts} from '../../../translations';
 import {DetailScreenNavigationProp} from '../Details';
+import {Coordinates} from '@entur/sdk';
+import {Position} from 'geojson';
 
 export type MapDetailRouteParams = {
-  legs: Leg[];
+  legs: MapLeg[];
+  fromPlace?: Coordinates | Position;
+  toPlace?: Coordinates | Position;
 };
 export type MapDetailRouteProp = RouteProp<DetailsStackParams, 'DetailsMap'>;
 type MapProps = {
@@ -33,11 +37,11 @@ export const TravelDetailsMap: React.FC<MapProps> = ({route, navigation}) => {
   const mapCameraRef = useRef<MapboxGL.Camera>(null);
   const mapViewRef = useRef<MapboxGL.MapView>(null);
   const {location: geolocation} = useGeolocationState();
-  const {legs} = route.params;
+  const {legs, toPlace, fromPlace} = route.params;
 
-  const features = useMemo(() => legsToMapLines(legs), [legs]);
-  const startPoint = pointOf(legs[0].fromPlace);
-  const endPoint = pointOf(legs[legs.length - 1].toPlace);
+  const features = useMemo(() => createMapLines(legs), [legs]);
+  // const startPoint = pointOf(legs[0].fromPlace);
+  // const endPoint = pointOf(legs[legs.length - 1].toPlace);
   const bounds = getMapBounds(features);
 
   async function zoomIn() {
@@ -76,16 +80,20 @@ export const TravelDetailsMap: React.FC<MapProps> = ({route, navigation}) => {
         />
         <MapboxGL.UserLocation showsUserHeadingIndicator />
         <MapRoute lines={features}></MapRoute>
-        <MapLabel
-          point={endPoint}
-          id={'end'}
-          text={t(MapTexts.endPoint.label)}
-        ></MapLabel>
-        <MapLabel
-          point={startPoint}
-          id={'start'}
-          text={t(MapTexts.startPoint.label)}
-        ></MapLabel>
+        {toPlace && (
+          <MapLabel
+            point={pointOf(toPlace)}
+            id={'end'}
+            text={t(MapTexts.endPoint.label)}
+          />
+        )}
+        {fromPlace && (
+          <MapLabel
+            point={pointOf(fromPlace)}
+            id={'start'}
+            text={t(MapTexts.startPoint.label)}
+          />
+        )}
       </MapboxGL.MapView>
       <View style={controlStyles.backArrowContainer}>
         <BackArrow
