+++ src/sdk.ts
@@ -1,6 +1,7 @@
 import {
   Authority,
   BikeRentalStation,
+  Coordinates,
   DestinationDisplay,
   InfoLink,
   Interchange,
@@ -215,7 +216,14 @@ export type CursorInput = {
 export type CursoredQuery<T> = CursorInput & T;
 
 export type MapLeg = {
+  mode?: LegMode;
+  faded?: boolean;
+  transportSubmode?: TransportSubmode;
   pointsOnLink: PointsOnLink;
-  mode: LegMode;
-  transportSubmode: TransportSubmode;
+};
+
+export type ServiceJourneyMapInfoData = {
+  mapLegs: MapLeg[];
+  start?: Coordinates;
+  stop?: Coordinates;
 };
